requirements.txt
aiogram>=3.3.0,<4.0.0
aiohttp>=3.9.5,<4.0.0
sqlalchemy>=2.0.30,<3.0.0
easyocr>=1.7.1,<2.0.0
Pillow>=10.3.0,<11.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
cachetools>=5.3.0,<6.0.0
